Move installation of cron jobs to the very end of the process
To avoid timing issues with the cron installation it must be moved to
the very end of the process. The following has been experienced:

The cron jobs get installed before the cleanup hook is run, which makes
vital file permissions and ownership changes after the installation. The
cron jobs depend on these changes but run before they occur, resulting
in failed scrip execution.