Change defaults for dirty copy questions and modify target handling
If a dirty copy is detected then the default value for the answer is now
N instead of Y.
Passing the deployment target changed from using a task with the name of
the target to a task called "target" followed by a colon and the name of
the target. This allows for dynamic adding of targets.

Example: Deploy to a target called xyz
   fab target:xyz depoy